+++ python/gilded_rose/__init__.py
@@ -1,5 +1,10 @@
 """Init for the gilded rose."""
 
+from gilded_rose import dtypes
+
+
+Item = dtypes.Item
+
 
 class GildedRose(object):
     def __init__(self, items):
@@ -37,13 +42,3 @@ class GildedRose(object):
                 else:
                     if item.quality < 50:
                         item.quality = item.quality + 1
-
-
-class Item:
-    def __init__(self, name, sell_in, quality):
-        self.name = name
-        self.sell_in = sell_in
-        self.quality = quality
-
-    def __repr__(self):
-        return "%s, %s, %s" % (self.name, self.sell_in, self.quality)
